+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 ### Fixed
-- Error on retrieving GPU information, caused by not properly understanding PNP Device ID. See issue #217
+- Missing icon for prompts
 
 ## [1.18.0] - 2024-06-06
 ### Added